Rename to `bingcd`
