+++ src/uint.rs
@@ -461,9 +461,9 @@ impl_uint_concat_split_mixed! {
     (U1024, [1, 2, 3, 4, 5, 6, 7, 9, 10, 11, 12, 13, 14, 15]),
 }
 
+mod bingcd;
 #[cfg(feature = "extra-sizes")]
 mod extra_sizes;
-mod new_gcd;
 
 #[cfg(test)]
 #[allow(clippy::unwrap_used)]
