+++ src/uint/bingcd/extension.rs
